feat: 实现剧情 VM 当前世界重载
1. 恢复opcode155的map22分支、deferred写入、同步reload与yield流程。
2. 保留完整tile dword、低16持久字段和陈旧空间绑定合同。
3. 补齐独立证据、工作包进度及定向/完整门验证。

+++ include/openswd3/world_map/legacy_maps_world_database.hpp
@@ -34,8 +34,8 @@ struct LegacyMapsWorldHeader {
 
 struct LegacyWorldLoadRequest {
     compat::u16 logical_map_id{};
-    compat::u16 tile_x{};
-    compat::u16 tile_y{};
+    compat::u32 tile_x{};
+    compat::u32 tile_y{};
     compat::u16 action_id{};
     compat::u16 base_variant{};
     compat::u16 variant_delta{};
